Further simplifications to save&restore snapshot view
