+++ app/save-and-restore/app/src/main/java/org/phoebus/applications/saveandrestore/ui/snapshot/CompareSnapshotsController.java
@@ -21,7 +21,7 @@ package org.phoebus.applications.saveandrestore.ui.snapshot;
 
 import javafx.fxml.FXML;
 
-public class CompareSnapshotsController extends RestoreSnapshotController{
+public class CompareSnapshotsController extends SnapshotController{
 
     public CompareSnapshotsController(SnapshotTab snapshotTab){
         super(snapshotTab);
