+++ app/save-and-restore/app/src/main/java/org/phoebus/applications/saveandrestore/ui/snapshot/SnapshotTableViewController.java
@@ -19,45 +19,53 @@
 
 package org.phoebus.applications.saveandrestore.ui.snapshot;
 
+import javafx.application.Platform;
+import javafx.beans.property.ObjectProperty;
 import javafx.beans.property.SimpleBooleanProperty;
 import javafx.collections.ObservableList;
+import javafx.event.ActionEvent;
 import javafx.fxml.FXML;
 import javafx.scene.control.*;
 import javafx.scene.control.cell.CheckBoxTableCell;
+import org.epics.vtype.VNumber;
+import org.epics.vtype.VNumberArray;
 import org.epics.vtype.VType;
 import org.phoebus.applications.saveandrestore.Messages;
-import org.phoebus.applications.saveandrestore.common.Utilities;
-import org.phoebus.applications.saveandrestore.common.VTypePair;
-import org.phoebus.applications.saveandrestore.model.Snapshot;
+import org.phoebus.applications.saveandrestore.Preferences;
+import org.phoebus.applications.saveandrestore.SafeMultiply;
+import org.phoebus.applications.saveandrestore.common.*;
+import org.phoebus.applications.saveandrestore.model.*;
+import org.phoebus.framework.jobs.JobManager;
 
+import java.text.SimpleDateFormat;
 import java.time.Instant;
-import java.util.List;
+import java.util.*;
+import java.util.concurrent.*;
+import java.util.function.Consumer;
+import java.util.logging.Level;
+import java.util.regex.Pattern;
+import java.util.stream.Collectors;
 
 
 public class SnapshotTableViewController extends BaseSnapshotTableViewController {
 
     @FXML
     private TooltipTableColumn<Instant> timeColumn;
-
     @FXML
     private TooltipTableColumn<VType> storedReadbackColumn;
-
     @FXML
     private TooltipTableColumn<VType> liveReadbackColumn;
-
     @FXML
     private TableColumn<TableEntry, ?> readbackColumn;
-
     private final SimpleBooleanProperty selectionInverted = new SimpleBooleanProperty(false);
-
     private final SimpleBooleanProperty showReadbacks = new SimpleBooleanProperty(false);
-    private boolean showDeltaPercentage;
+    private final SimpleBooleanProperty showDeltaPercentage = new SimpleBooleanProperty(false);
+    private final SimpleBooleanProperty hideEqualItems = new SimpleBooleanProperty(false);
 
     @FXML
     public void initialize() {
 
         super.initialize();
-
 
         selectedColumn.setCellFactory(column -> new SelectionCell());
 
@@ -86,17 +94,22 @@ public class SnapshotTableViewController extends BaseSnapshotTableViewController
         timeColumn.setCellFactory(c -> new SnapshotTableViewController.TimestampTableCell());
 
         deltaColumn.setCellValueFactory(e -> e.getValue().valueProperty());
-        deltaColumn.setCellFactory(e -> {
+        deltaColumn.setCellFactory(e -> new VDeltaCellEditor<>());
+
+        showDeltaPercentage.addListener((ob, o, n) -> deltaColumn.setCellFactory(e -> {
             VDeltaCellEditor<VTypePair> vDeltaCellEditor = new VDeltaCellEditor<>();
-            if (showDeltaPercentage) {
-                vDeltaCellEditor.setShowDeltaPercentage();
-            }
+            vDeltaCellEditor.setShowDeltaPercentage(n);
             return vDeltaCellEditor;
-        });
+        }));
+
         deltaColumn.setComparator((pair1, pair2) -> {
             Utilities.VTypeComparison vtc1 = Utilities.valueToCompareString(pair1.value, pair1.base, pair1.threshold);
             Utilities.VTypeComparison vtc2 = Utilities.valueToCompareString(pair2.value, pair2.base, pair2.threshold);
             return Double.compare(vtc1.getAbsoluteDelta(), vtc2.getAbsoluteDelta());
+        });
+
+        hideEqualItems.addListener((ob, o, n) -> {
+
         });
 
         liveReadbackColumn.setCellFactory(e -> new VTypeCellEditor<>());
@@ -108,44 +121,7 @@ public class SnapshotTableViewController extends BaseSnapshotTableViewController
         this.snapshotController = snapshotController;
     }
 
-    public void updateTable(List<TableEntry> entries, List<Snapshot> snapshots, boolean showLiveReadback, boolean showDeltaPercentage) {
-        uiSnapshots.clear();
-        // we should always know if we are showing the stored readback or not, to properly extract the selection
-        this.showReadbacks.set(showLiveReadback);
-        this.showDeltaPercentage = showDeltaPercentage;
-        uiSnapshots.addAll(snapshots);
-        updateTable(entries);
-    }
 
-    /**
-     * Sets new table entries for this table, but do not change the structure of the table.
-     *
-     * @param entries the entries to set
-     */
-    public void updateTable(List<TableEntry> entries) {
-        final ObservableList<TableEntry> items = snapshotTableView.getItems();
-        final boolean notHide = !snapshotController.isHideEqualItems();
-        items.clear();
-        entries.forEach(e -> {
-            // there is no harm if this is executed more than once, because only one line is allowed for these
-            // two properties (see SingleListenerBooleanProperty for more details)
-            e.liveStoredEqualProperty().addListener((a, o, n) -> {
-                if (snapshotController.isHideEqualItems()) {
-                    if (n) {
-                        snapshotTableView.getItems().remove(e);
-                    } else {
-                        snapshotTableView.getItems().add(e);
-                    }
-                }
-            });
-            if (notHide || !e.liveStoredEqualProperty().get()) {
-                items.add(e);
-            }
-            e.selectedProperty().addListener((ob, o, n) -> {
-                System.out.println(e.pvNameProperty().get() + " " + n);
-            });
-        });
-    }
 
     /**
      * {@link CheckBoxTableCell} handling the selection checkboxes for each row in the table.
@@ -163,5 +139,234 @@ public class SnapshotTableViewController extends BaseSnapshotTableViewController
                 setGraphic(null);
             }
         }
+    }
+
+    public void takeSnapshot(Consumer<Snapshot> consumer) {
+        List<SnapshotItem> entries = new ArrayList<>();
+        readAll(list ->
+                Platform.runLater(() -> {
+                    tableEntryItems.clear();
+                    entries.addAll(list);
+                    Snapshot snapshot = new Snapshot();
+                    snapshot.setSnapshotNode(Node.builder().nodeType(NodeType.SNAPSHOT).build());
+                    SnapshotData snapshotData = new SnapshotData();
+                    snapshotData.setSnapshotItems(entries);
+                    snapshot.setSnapshotData(snapshotData);
+                    List<TableEntry> tableEntries = createTableEntries(snapshot);
+                    updateTable(tableEntries);
+                    if (!Preferences.default_snapshot_name_date_format.equals("")) {
+                        SimpleDateFormat formatter = new SimpleDateFormat(Preferences.default_snapshot_name_date_format);
+                        snapshot.getSnapshotNode().setName(formatter.format(new Date()));
+                    }
+                    consumer.accept(snapshot);
+                })
+        );
+    }
+
+    /**
+     * Reads all PVs using a thread pool. All reads are asynchronous, waiting at most the amount of time
+     * configured through a preference setting.
+     *
+     * @param completion Callback receiving a list of {@link SnapshotItem}s where values for PVs that could
+     *                   not be read are set to {@link VDisconnectedData#INSTANCE}.
+     */
+    private void readAll(Consumer<List<SnapshotItem>> completion) {
+        ExecutorService executorService = Executors.newFixedThreadPool(10);
+        SnapshotItem[] snapshotEntries = new SnapshotItem[tableEntryItems.values().size()];
+        JobManager.schedule("Take snapshot", monitor -> {
+            final CountDownLatch countDownLatch = new CountDownLatch(tableEntryItems.values().size());
+            for (TableEntry t : tableEntryItems.values()) {
+                // Submit read request only if job has not been cancelled
+                executorService.submit(() -> {
+                    SaveAndRestorePV pv = pvs.get(getPVKey(t.pvNameProperty().get(), t.readOnlyProperty().get()));
+                    VType value = VDisconnectedData.INSTANCE;
+                    try {
+                        value = pv.getPv().asyncRead().get(Preferences.readTimeout, TimeUnit.MILLISECONDS);
+                    } catch (Exception e) {
+                        LOGGER.log(Level.WARNING, "Failed to read PV " + pv.getPvName(), e);
+                    }
+                    VType readBackValue = VDisconnectedData.INSTANCE;
+                    if (pv.getReadbackPv() != null && !pv.getReadbackValue().equals(VDisconnectedData.INSTANCE)) {
+                        try {
+                            readBackValue = pv.getReadbackPv().asyncRead().get(Preferences.readTimeout, TimeUnit.MILLISECONDS);
+                        } catch (Exception e) {
+                            LOGGER.log(Level.WARNING, "Failed to read read-back PV " + pv.getReadbackPvName(), e);
+                        }
+                    }
+                    SnapshotItem snapshotItem = new SnapshotItem();
+                    snapshotItem.setConfigPv(t.getConfigPv());
+                    snapshotItem.setValue(value);
+                    snapshotItem.setReadbackValue(readBackValue);
+
+                    snapshotEntries[t.idProperty().get() - 1] = snapshotItem;
+                    countDownLatch.countDown();
+                });
+            }
+            countDownLatch.await();
+            completion.accept(Arrays.asList(snapshotEntries));
+            executorService.shutdown();
+        });
+    }
+
+    public void updateThreshold(Snapshot snapshot, double threshold) {
+        snapshot.getSnapshotData().getSnapshotItems().forEach(item -> {
+            VType vtype = item.getValue();
+            VNumber diffVType;
+
+            double ratio = threshold / 100;
+
+            TableEntry tableEntry = tableEntryItems.get(getPVKey(item.getConfigPv().getPvName(), item.getConfigPv().isReadOnly()));
+            if (tableEntry == null) {
+                tableEntry = tableEntryItems.get(getPVKey(item.getConfigPv().getPvName(), !item.getConfigPv().isReadOnly()));
+            }
+
+            if (!item.getConfigPv().equals(tableEntry.getConfigPv())) {
+                return;
+            }
+
+            if (vtype instanceof VNumber) {
+                diffVType = SafeMultiply.multiply((VNumber) vtype, ratio);
+                VNumber vNumber = diffVType;
+                boolean isNegative = vNumber.getValue().doubleValue() < 0;
+
+                tableEntry.setThreshold(Optional.of(new Threshold<>(isNegative ? SafeMultiply.multiply(vNumber.getValue(), -1.0) : vNumber.getValue())));
+            }
+        });
+    }
+
+    public void updateSnapshotValues(Snapshot snapshot, double multiplier) {
+        snapshot.getSnapshotData().getSnapshotItems()
+                .forEach(item -> {
+                    TableEntry tableEntry = tableEntryItems.get(getPVKey(item.getConfigPv().getPvName(), item.getConfigPv().isReadOnly()));
+                    VType vtype = tableEntry.getStoredValue().get();
+                    VType newVType;
+
+                    if (vtype instanceof VNumber) {
+                        newVType = SafeMultiply.multiply((VNumber) vtype, multiplier);
+                    } else if (vtype instanceof VNumberArray) {
+                        newVType = SafeMultiply.multiply((VNumberArray) vtype, multiplier);
+                    } else {
+                        return;
+                    }
+
+                    item.setValue(newVType);
+
+                    tableEntry.storedValueProperty().set(newVType);
+
+                    ObjectProperty<VTypePair> value = tableEntry.valueProperty();
+                    value.setValue(new VTypePair(value.get().base, newVType, value.get().threshold));
+                });
+    }
+
+    public void applyFilter(String filterText, boolean preserveSelection, List<List<Pattern>> regexPatterns) {
+        if (filterText.isEmpty()) {
+            List<TableEntry> arrayList = tableEntryItems.values().stream()
+                    .peek(item -> {
+                        if (!preserveSelection) {
+                            if (!item.readOnlyProperty().get()) {
+                                item.selectedProperty().set(true);
+                            }
+                        }
+                    }).collect(Collectors.toList());
+
+            Platform.runLater(() -> updateTable(arrayList));
+            return;
+        }
+
+        List<TableEntry> filteredEntries = tableEntryItems.values().stream()
+                .filter(item -> {
+                    boolean matchEither = false;
+                    for (List<Pattern> andPatternList : regexPatterns) {
+                        boolean matchAnd = true;
+                        for (Pattern pattern : andPatternList) {
+                            matchAnd &= pattern.matcher(item.pvNameProperty().get()).find();
+                        }
+
+                        matchEither |= matchAnd;
+                    }
+
+                    if (!preserveSelection) {
+                        item.selectedProperty().setValue(matchEither);
+                    } else {
+                        matchEither |= item.selectedProperty().get();
+                    }
+
+                    return matchEither;
+                }).collect(Collectors.toList());
+
+        Platform.runLater(() -> updateTable(filteredEntries));
+    }
+
+    public void applyPreserveSelection(boolean preserve) {
+        if (preserve) {
+            boolean allSelected = tableEntryItems.values().stream().allMatch(item -> item.selectedProperty().get());
+
+            if (allSelected) {
+                tableEntryItems.values()
+                        .forEach(item -> item.selectedProperty().set(false));
+            }
+        }
+    }
+
+    public void showReadback(boolean showLiveReadback) {
+        this.showReadbacks.set(showLiveReadback);
+    }
+
+    public void hideEqualItems(boolean hide) {
+        ArrayList<TableEntry> arrayList = new ArrayList<>(tableEntryItems.values());
+        Platform.runLater(() -> updateTable(arrayList));
+    }
+
+    public void restore(Snapshot snapshot, Consumer<List<String>> completion) {
+        new Thread(() -> {
+            List<String> restoreFailedPVNames = new ArrayList<>();
+            CountDownLatch countDownLatch = new CountDownLatch(snapshot.getSnapshotData().getSnapshotItems().size());
+            snapshot.getSnapshotData().getSnapshotItems()
+                    .forEach(e -> pvs.get(getPVKey(e.getConfigPv().getPvName(), e.getConfigPv().isReadOnly())).setCountDownLatch(countDownLatch));
+
+            for (SnapshotItem entry : snapshot.getSnapshotData().getSnapshotItems()) {
+                TableEntry e = tableEntryItems.get(getPVKey(entry.getConfigPv().getPvName(), entry.getConfigPv().isReadOnly()));
+
+                boolean restorable = e.selectedProperty().get() && !e.readOnlyProperty().get() &&
+                        !entry.getValue().equals(VNoData.INSTANCE);
+
+                if (restorable) {
+                    final SaveAndRestorePV pv = pvs.get(getPVKey(e.pvNameProperty().get(), e.readOnlyProperty().get()));
+                    if (entry.getValue() != null) {
+                        try {
+                            pv.getPv().write(Utilities.toRawValue(entry.getValue()));
+                        } catch (Exception writeException) {
+                            restoreFailedPVNames.add(entry.getConfigPv().getPvName());
+                        } finally {
+                            pv.countDown();
+                        }
+                    }
+                } else {
+                    countDownLatch.countDown();
+                }
+            }
+
+            try {
+                countDownLatch.await();
+            } catch (InterruptedException e) {
+                LOGGER.log(Level.INFO, "Encountered InterruptedException", e);
+            }
+
+            if (restoreFailedPVNames.isEmpty()) {
+                LOGGER.log(Level.FINE, "Restored snapshot {0}", snapshot.getSnapshotNode().getName());
+            } else {
+                Collections.sort(restoreFailedPVNames);
+                StringBuilder sb = new StringBuilder(restoreFailedPVNames.size() * 200);
+                restoreFailedPVNames.forEach(e -> sb.append(e).append('\n'));
+                LOGGER.log(Level.WARNING,
+                        "Not all PVs could be restored for {0}: {1}. The following errors occurred:\n{2}",
+                        new Object[]{snapshot.getSnapshotNode().getName(), snapshot.getSnapshotNode(), sb.toString()});
+            }
+            completion.accept(restoreFailedPVNames);
+        }).start();
+    }
+
+    public void setShowDeltaPercentage(boolean showDeltaPercentage){
+        this.showDeltaPercentage.set(showDeltaPercentage);
     }
 }
