+++ app/save-and-restore/app/src/main/resources/org/phoebus/applications/saveandrestore/ui/snapshot/SnapshotControlsView.fxml
@@ -31,7 +31,7 @@
             </tooltip>
         </ToggleButton>
         <Pane HBox.hgrow="ALWAYS" />
-        <ToggleButton fx:id="showHideDeltaPercentageButton" mnemonicParsing="false">
+        <ToggleButton fx:id="showDeltaPercentageButton" mnemonicParsing="false">
             <tooltip>
                 <Tooltip text="%toolTipShowHideDeltaPercentageToggleButton" />
             </tooltip>
