+++ app/save-and-restore/app/src/main/resources/save-and-restore-style.css
@@ -115,6 +115,12 @@
     -fx-font-family: "Liberation Mono";
 }
 
+.id-column.table-cell{
+    -fx-font-family: "Liberation Mono";
+    -fx-alignment: CENTER-RIGHT;
+    -fx-padding: 0 4 2 0;
+}
+
 .filter-match {
   -fx-background-color: rgb(242, 242, 242);
 }
